+++ news/builders_folder.rst
@@ -1,0 +1,13 @@
+**Added:** None
+
+**Changed:**
+
+* moved builders into ``builders`` folder
+
+**Deprecated:** None
+
+**Removed:** None
+
+**Fixed:** None
+
+**Security:** None
