+++ regolith/builder.py
@@ -1,10 +1,10 @@
 """Generic builder."""
 
-from regolith.cvbuilder import CVBuilder
-from regolith.htmlbuilder import HtmlBuilder
-from regolith.publistbuilder import PubListBuilder
-from regolith.gradebuilder import GradeReportBuilder
-from regolith.resumebuilder import ResumeBuilder
+from regolith.builders.cvbuilder import CVBuilder
+from regolith.builders.htmlbuilder import HtmlBuilder
+from regolith.builders.publistbuilder import PubListBuilder
+from regolith.builders.gradebuilder import GradeReportBuilder
+from regolith.builders.resumebuilder import ResumeBuilder
 
 
 BUILDERS = {
