+++ regolith/builders/cvbuilder.py
@@ -1,11 +1,10 @@
 """Builder for CVs."""
 
-from regolith.basebuilder import LatexBuilderBase
+from regolith.builders.basebuilder import LatexBuilderBase
 from regolith.sorters import ene_date_key, position_key
-from regolith.tools import (all_docs_from_collection, month_and_year,
-                            filter_publications, filter_projects,
+from regolith.tools import (all_docs_from_collection, filter_publications, filter_projects,
                             filter_grants, awards_grants_honors,
-                            latex_safe, make_bibtex_file)
+                            make_bibtex_file)
 
 
 class CVBuilder(LatexBuilderBase):
